fix: remove maxAge from session cookie
Setting maxAge: 0 will automatically expire the cookie as soon as it's sent. This should not be the default value. Otherwise, the developer will have to override this on every set-cookie header. Leaving it undefined ensures the cookie will live for the life of the browser session.

See https://github.com/remix-run/blues-stack/pull/85